tests: move CorrelatedStack tests to test_imaging_camera directory
* split tests into multiple files
* add conftest
* run black
